feat(project): 新增项目与用户组管理功能

+++ frontend/src/views/setting/user-group/components/UserTable.vue
@@ -9,9 +9,13 @@ import type {SimpleUserInfo} from "/@/api/modules/user/types.ts";
 import BasePagination from "/@/components/BasePagination.vue";
 import RemoveButton from "/@/components/RemoveButton.vue";
 import {AuthScopeEnum, type AuthScopeEnumType} from "/@/enums/common-enum.ts";
+import {hasAnyPermission} from "/@/utils/permission.ts";
 
-const {current = {code: '', id: ''}} = defineProps<{
-  current: CurrentUserGroupItem
+const {current = {code: '', id: ''}, readPermission = [], updatePermission = []} = defineProps<{
+  current: CurrentUserGroupItem;
+  deletePermission?: string[];
+  readPermission?: string[];
+  updatePermission?: string[];
 }>()
 const systemType = inject<AuthScopeEnumType>('systemType');
 const currentKeyword = defineModel<string>('keyword')
@@ -37,26 +41,38 @@ const columns: DataTableColumns<SimpleUserInfo> = [
   {
     title: '操作', key: 'actions', fixed: 'right', width: 100,
     render(record) {
-      return h(RemoveButton, {
-        title: `确认移除 ${record.name} 这个用户吗？`, subTitleTip: '移除后，将失去用户组权限',
-        disabled: systemType === AuthScopeEnum.SYSTEM && record.name === 'admin',
-        onConfirm: () => handleRemove(record)
-      }, {})
+      if (hasAnyPermission(updatePermission)) {
+        return h(RemoveButton, {
+          title: `确认移除 ${record.name} 这个用户吗？`, subTitleTip: '移除后，将失去用户组权限',
+          disabled: systemType === AuthScopeEnum.SYSTEM && record.name === 'admin',
+          onConfirm: () => handleRemove(record)
+        }, {})
+      }
     }
 
   }
 ]
 const {send: deleteUserFromUserGroup} = useRequest((id) => userGroupApi.deleteUserFromUserGroup(id), {immediate: false})
 const handleRemove = (record: SimpleUserInfo) => {
-  if (systemType === AuthScopeEnum.SYSTEM) {
-    deleteUserFromUserGroup(record.id).then(() => {
-      fetchData();
-    })
-  }
+  handlePermission(updatePermission, () => {
+    if (systemType === AuthScopeEnum.SYSTEM) {
+      deleteUserFromUserGroup(record.id).then(() => {
+        fetchData();
+      })
+    }
+  })
 }
+const handlePermission = (permission: string[], cb: () => void) => {
+  if (!hasAnyPermission(permission)) {
+    return false;
+  }
+  cb();
+};
 watchEffect(() => {
   if (current.id) {
-    fetchData();
+    handlePermission(readPermission, () => {
+      fetchData();
+    })
   }
 })
 defineExpose({fetchData})
